Better error handling in authentication services and refactored some of the tests.
